# docs/implementation/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pongdang API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/tides/events": {
      "get": {
        "tags": [
          "tides"
        ],
        "summary": "Events",
        "operationId": "events_api_data_tides_events_get",
        "parameters": [
          {
            "name": "spot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Spot Id"
            }
          },
          {
            "name": "activity",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "swim",
                "surf",
                "relax",
                "mudflat",
                "onsen",
                "rafting"
              ],
              "type": "string",
              "title": "Activity"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Until"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "general",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "observation",
                    "forecast"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "reference_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reference At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TideEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/tides/windows": {
      "get": {
        "tags": [
          "tides"
        ],
        "summary": "Windows",
        "operationId": "windows_api_data_tides_windows_get",
        "parameters": [
          {
            "name": "spot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Spot Id"
            }
          },
          {
            "name": "activity",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "swim",
                "surf",
                "relax",
                "mudflat",
                "onsen",
                "rafting"
              ],
              "type": "string",
              "title": "Activity"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Until"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "general",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "observation",
                    "forecast"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WindowEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/quality/comparisons": {
      "get": {
        "tags": [
          "water-quality"
        ],
        "summary": "Comparisons",
        "operationId": "comparisons_api_data_quality_comparisons_get",
        "parameters": [
          {
            "name": "spot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Spot Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/quality/observations": {
      "get": {
        "tags": [
          "water-quality"
        ],
        "summary": "Observations",
        "operationId": "observations_api_data_quality_observations_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerObservationEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "water-quality"
        ],
        "summary": "Observation",
        "operationId": "observation_api_data_quality_observations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComparisonEnvelope": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "water-quality.v1",
            "title": "Contract Version"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ComparisonRow"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rows"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "analysis_pending"
            ],
            "title": "Status"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Reason Codes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contract_version",
          "rows",
          "total",
          "page",
          "page_size",
          "as_of",
          "queried_at",
          "status",
          "reason_codes"
        ],
        "title": "ComparisonEnvelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObservationInput": {
        "properties": {
          "review_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Review Id"
          },
          "revision": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Revision"
          },
          "spot_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spot Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "review",
              "field_observation"
            ],
            "title": "Kind",
            "default": "review"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "observed_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Until"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Asia/Seoul"
          },
          "spatial_relation": {
            "type": "string",
            "enum": [
              "at_spot",
              "nearby",
              "unknown"
            ],
            "title": "Spatial Relation"
          },
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text"
          },
          "source_record_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Record Id"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/Measurement"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Measurements",
            "default": []
          },
          "retracted": {
            "type": "boolean",
            "title": "Retracted",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "review_id",
          "revision",
          "spot_id",
          "observed_at",
          "spatial_relation",
          "text"
        ],
        "title": "ObservationInput"
      },
      "ObservationReceipt": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id"
          },
          "review_id": {
            "type": "string",
            "title": "Review Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "created": {
            "type": "boolean",
            "title": "Created"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/Signal"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Signals"
          },
          "analysis_version": {
            "type": "string",
            "title": "Analysis Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "retracted",
              "analysis_pending"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evidence_id",
          "review_id",
          "revision",
          "created",
          "received_at",
          "signals",
          "analysis_version",
          "status"
        ],
        "title": "ObservationReceipt"
      },
...
      "OwnerObservation": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "input": {
            "$ref": "#/components/schemas/ObservationInput"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/Signal"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Signals"
          },
          "analysis_version": {
            "type": "string",
            "title": "Analysis Version"
          },
          "duplicate_fingerprint": {
            "type": "string",
            "title": "Duplicate Fingerprint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evidence_id",
          "received_at",
          "input",
          "signals",
          "analysis_version",
          "duplicate_fingerprint"
        ],
        "title": "OwnerObservation"
      },
      "OwnerObservationEnvelope": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/OwnerObservation"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rows"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rows",
          "total",
          "page",
          "page_size"
        ],
        "title": "OwnerObservationEnvelope"
      },
...
      "TideEnvelope": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "tide-timer.v1",
            "title": "Contract Version"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "reference_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reference At"
          },
          "page": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/TideEvent"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rows"
          },
          "next_high": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TideEvent"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_low": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TideEvent"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "no_forecast_data",
              "outside_forecast_horizon",
              "missing_within_horizon"
            ],
            "title": "Status"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Reason Codes"
          },
          "next_event_scope": {
            "type": "string",
            "const": "requested_interval",
            "title": "Next Event Scope"
          },
          "horizon_start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon Start At"
          },
          "horizon_end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon End At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contract_version",
          "as_of",
          "queried_at",
          "reference_at",
          "page",
          "page_size",
          "total",
          "rows",
          "next_high",
          "next_low",
          "status",
          "reason_codes",
          "next_event_scope",
          "horizon_start_at",
          "horizon_end_at"
        ],
        "title": "TideEnvelope"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WindowEnvelope": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "tide-operating-window.v1",
            "title": "Contract Version"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "page": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PublicOperatingWindow"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rows"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "no_official_operating_window"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contract_version",
          "as_of",
          "queried_at",
          "page",
          "page_size",
          "total",
          "rows",
          "status"
        ],
        "title": "WindowEnvelope"
      }
    }
  }
}